+++ src/Stub/AddEventListenerOptions.php
@@ -8,7 +8,7 @@ namespace Wikimedia\IDLeDOM\Stub;
 use Wikimedia\IDLeDOM\AbortSignal;
 
 trait AddEventListenerOptions {
-	use \Wikimedia\IDLeDOM\Stub\EventListenerOptions;
+	// use \Wikimedia\IDLeDOM\Stub\EventListenerOptions;
 
 	// Underscore is used to avoid conflicts with DOM-reserved names
 	// phpcs:disable PSR2.Methods.MethodDeclaration.Underscore