+++ src/Stub/Document.php
@@ -24,11 +24,12 @@ use Wikimedia\IDLeDOM\Text;
 use Wikimedia\IDLeDOM\TreeWalker;
 
 trait Document {
-	use \Wikimedia\IDLeDOM\Stub\Node;
-	use \Wikimedia\IDLeDOM\Stub\DocumentOrShadowRoot;
-	use \Wikimedia\IDLeDOM\Stub\NonElementParentNode;
-	use \Wikimedia\IDLeDOM\Stub\ParentNode;
-	use \Wikimedia\IDLeDOM\Stub\XPathEvaluatorBase;
+	// use \Wikimedia\IDLeDOM\Stub\Node;
+	// use \Wikimedia\IDLeDOM\Stub\EventTarget;
+	// use \Wikimedia\IDLeDOM\Stub\DocumentOrShadowRoot;
+	// use \Wikimedia\IDLeDOM\Stub\NonElementParentNode;
+	// use \Wikimedia\IDLeDOM\Stub\ParentNode;
+	// use \Wikimedia\IDLeDOM\Stub\XPathEvaluatorBase;
 
 	// Underscore is used to avoid conflicts with DOM-reserved names
 	// phpcs:disable PSR2.Methods.MethodDeclaration.Underscore