+++ src/Stub/ProcessingInstruction.php
@@ -6,7 +6,11 @@
 namespace Wikimedia\IDLeDOM\Stub;
 
 trait ProcessingInstruction {
-	use \Wikimedia\IDLeDOM\Stub\CharacterData;
+	// use \Wikimedia\IDLeDOM\Stub\CharacterData;
+	// use \Wikimedia\IDLeDOM\Stub\Node;
+	// use \Wikimedia\IDLeDOM\Stub\EventTarget;
+	// use \Wikimedia\IDLeDOM\Stub\ChildNode;
+	// use \Wikimedia\IDLeDOM\Stub\NonDocumentTypeChildNode;
 
 	// Underscore is used to avoid conflicts with DOM-reserved names
 	// phpcs:disable PSR2.Methods.MethodDeclaration.Underscore