+++ src/Stub/StaticRange.php
@@ -6,7 +6,7 @@
 namespace Wikimedia\IDLeDOM\Stub;
 
 trait StaticRange {
-	use \Wikimedia\IDLeDOM\Stub\AbstractRange;
+	// use \Wikimedia\IDLeDOM\Stub\AbstractRange;
 
 	// Underscore is used to avoid conflicts with DOM-reserved names
 	// phpcs:disable PSR2.Methods.MethodDeclaration.Underscore