+++ src/Stub/Text.php
@@ -6,8 +6,12 @@
 namespace Wikimedia\IDLeDOM\Stub;
 
 trait Text {
-	use \Wikimedia\IDLeDOM\Stub\CharacterData;
-	use \Wikimedia\IDLeDOM\Stub\Slottable;
+	// use \Wikimedia\IDLeDOM\Stub\CharacterData;
+	// use \Wikimedia\IDLeDOM\Stub\Node;
+	// use \Wikimedia\IDLeDOM\Stub\EventTarget;
+	// use \Wikimedia\IDLeDOM\Stub\ChildNode;
+	// use \Wikimedia\IDLeDOM\Stub\NonDocumentTypeChildNode;
+	// use \Wikimedia\IDLeDOM\Stub\Slottable;
 
 	// Underscore is used to avoid conflicts with DOM-reserved names
 	// phpcs:disable PSR2.Methods.MethodDeclaration.Underscore