+++ src/Stub/XMLDocument.php
@@ -6,7 +6,13 @@
 namespace Wikimedia\IDLeDOM\Stub;
 
 trait XMLDocument {
-	use \Wikimedia\IDLeDOM\Stub\Document;
+	// use \Wikimedia\IDLeDOM\Stub\Document;
+	// use \Wikimedia\IDLeDOM\Stub\Node;
+	// use \Wikimedia\IDLeDOM\Stub\EventTarget;
+	// use \Wikimedia\IDLeDOM\Stub\DocumentOrShadowRoot;
+	// use \Wikimedia\IDLeDOM\Stub\NonElementParentNode;
+	// use \Wikimedia\IDLeDOM\Stub\ParentNode;
+	// use \Wikimedia\IDLeDOM\Stub\XPathEvaluatorBase;
 
 	// Underscore is used to avoid conflicts with DOM-reserved names
 	// phpcs:disable PSR2.Methods.MethodDeclaration.Underscore