+++ src/Stub/XPathEvaluator.php
@@ -6,7 +6,7 @@
 namespace Wikimedia\IDLeDOM\Stub;
 
 trait XPathEvaluator {
-	use \Wikimedia\IDLeDOM\Stub\XPathEvaluatorBase;
+	// use \Wikimedia\IDLeDOM\Stub\XPathEvaluatorBase;
 
 	// Underscore is used to avoid conflicts with DOM-reserved names
 	// phpcs:disable PSR2.Methods.MethodDeclaration.Underscore